# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API-Notebook",
        "version": "1.0"
    },
    "paths": {
        "/api/": {},
        "/api/v1/notebook": {
            "get": {
                "tags": [
                    "CRUD записная книжка"
                ],
                "summary": "Список записей",
                "operationId": "ddadc481df9b4820c25eace6094daed0",
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Пример имени"
                                                    },
                                                    "company": {
                                                        "type": "string",
                                                        "example": "Пример компании"
                                                    },
                                                    "phone": {
                                                        "type": "string",
                                                        "example": "Пример телефона"
                                                    },
                                                    "email": {
                                                        "type": "email",
                                                        "example": "test@test.ru"
                                                    },
                                                    "birthdate": {
                                                        "type": "string",
                                                        "format": "date",
                                                        "example": "01.01.2001"
                                                    },
                                                    "photo": {
                                                        "type": "string",
                                                        "example": "Пример пути до фото"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "CRUD записная книжка"
                ],
                "summary": "Создание записи",
                "operationId": "38b5c77e4129659b839f7feb26c20e9d",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Пример имени"
                                            },
                                            "company": {
                                                "type": "string",
                                                "example": "Пример компании"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "Пример телефона"
                                            },
                                            "email": {
                                                "type": "email",
                                                "example": "test@test.ru"
                                            },
                                            "birthdate": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "01.01.2001"
                                            },
                                            "photo": {
                                                "type": "string",
                                                "example": "Пример пути до фото"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Пример имени"
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Пример компании"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "Пример телефона"
                                                },
                                                "email": {
                                                    "type": "email",
                                                    "example": "test@test.ru"
                                                },
                                                "birthdate": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "01.01.2001"
                                                },
                                                "photo": {
                                                    "type": "string",
                                                    "example": "Пример пути до фото"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/notebook/{notebook}": {
            "get": {
                "tags": [
                    "CRUD записная книжка"
                ],
                "summary": "Конкретная запись",
                "operationId": "8e729330a735a6f3cedcff3103605d9b",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Пример имени"
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Пример компании"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "Пример телефона"
                                                },
                                                "email": {
                                                    "type": "email",
                                                    "example": "test@test.ru"
                                                },
                                                "birthdate": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "01.01.2001"
                                                },
                                                "photo": {
                                                    "type": "string",
                                                    "example": "Пример пути до фото"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "CRUD записная книжка"
                ],
                "summary": "Удаление записи",
                "operationId": "b000f9d57073abe2bfc9d4e10a90be01",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "done"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "CRUD записная книжка"
                ],
                "summary": "Обновление записи",
                "operationId": "31f86602be4efb61ec9f26cf3b3aca45",
                "parameters": [
                    {
                        "name": "notebook",
                        "in": "path",
                        "description": "ID записи",
                        "required": true,
                        "example": 3
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "allOf": [
                                    {
                                        "properties": {
                                            "name": {
                                                "type": "string",
                                                "example": "Пример имени"
                                            },
                                            "company": {
                                                "type": "string",
                                                "example": "Пример компании"
                                            },
                                            "phone": {
                                                "type": "string",
                                                "example": "Пример телефона"
                                            },
                                            "email": {
                                                "type": "email",
                                                "example": "test@test.ru"
                                            },
                                            "birthdate": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "01.01.2001"
                                            },
                                            "photo": {
                                                "type": "string",
                                                "example": "Пример пути до фото"
                                            }
                                        },
                                        "type": "object"
                                    }
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Пример имени"
                                                },
                                                "company": {
                                                    "type": "string",
                                                    "example": "Пример компании"
                                                },
                                                "phone": {
                                                    "type": "string",
                                                    "example": "Пример телефона"
                                                },
                                                "email": {
                                                    "type": "email",
                                                    "example": "test@test.ru"
                                                },
                                                "birthdate": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "01.01.2001"
                                                },
                                                "photo": {
                                                    "type": "string",
                                                    "example": "Пример пути до фото"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}